requirements.txt
httpx
matrix-nio
markdown
fastapi
uvicorn
markdown
fastapi
uvicorn
pydantic
requests
python-dotenv
python-multipart
